Polish
Plunger goes back to start position, changes in board, fixed lives counter to properly go down by 1. TODO: sounds

+++ Assets/Scripts/Net.cs
@@ -7,7 +7,7 @@ public class Net : MonoBehaviour
 {
     public Rigidbody r_ball;
     public GameObject ballspawn;
-    static int lives = 3; //can use this later to implement a lives system and end screen
+    static float lives = 3; //can use this later to implement a lives system and end screen
     public TextMeshProUGUI livesText;
